+++ components/props/lots_of_greetings.js
