+++ components/state/blink.js
