+++ components/style/lots_of_styles.js
