+++ index.ios.js
@@ -5,24 +5,30 @@
  */
 
 import React, { Component } from 'react';
-import {
-  AppRegistry,
-  Text,
-} from 'react-native';
-import BlinkApp from './components/blink_app';
-import LotsOfStyles from './components/lots_of_styles';
-import FixedDimensionsBasics from './components/fixed_dimensions_basics';
-import FlexDimensionsBasics from './components/flex_dimensions_basics';
+import { AppRegistry } from 'react-native';
+
+/* https://facebook.github.io/react-native/docs/props.html
+import Bananas from './components/props/bananas';
+import LotsOfGreetings from './components/props/lots_of_greetings';
+*/
+
+/* https://facebook.github.io/react-native/docs/state.html
+import BlinkApp from './components/state/blink_app';
+*/
+
+/* https://facebook.github.io/react-native/docs/style.html
+import LotsOfStyles from './components/style/lots_of_styles';
+*/
+
+/* https://facebook.github.io/react-native/docs/height-and-width.html
+import FixedDimensionsBasics from './components/height_and_width/fixed_dimensions_basics';
+import FlexDimensionsBasics from './components/height_and_width/flex_dimensions_basics';
+*/
+
+/* https://facebook.github.io/react-native/docs/flexbox.html
 import FlexDirectionBasics from './components/flexbox/flex_direction_basics';
 import JustifyContentBasics from './components/flexbox/justify_content_basics';
+*/
 import AlignItemsBasics from './components/flexbox/align_items_basics';
-
-export default class TestApp extends Component {
-  render() {
-    return (
-      <Text>Hello World!</Text>
-    );
-  }
-}
 
 AppRegistry.registerComponent('TestApp', () => AlignItemsBasics);
